improve behavior
